// code-review/ex3-nest-angular/backend/src/domain/enums/notification-type.enum.ts
export enum NotificationType {
  BID_PLACED = 'BID_PLACED',
  OUTBID = 'OUTBID',
  AUCTION_WON = 'AUCTION_WON',
  AUCTION_ENDED = 'AUCTION_ENDED',
  AUCTION_STARTING = 'AUCTION_STARTING',
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
  ITEM_APPROVED = 'ITEM_APPROVED',
  ITEM_REJECTED = 'ITEM_REJECTED',
}